Add server configuration with JSON file
- Try each consecutive config in an array to attempt sending the email.
- The configuration object is based on the options object for the
nodemailer API.
- Make config.json untracked so the username/password aren't public
:laughing:
