+++ .gitignore
@@ -1,3 +1,2 @@
 node_modules/
-dist/
-.DS_Store
+server/config.json
